Temporarily remove the actual TA's
In the long run the TA's should be here, but since we are in the middle
of refactoring and debating how to work with the TA's
(optee_os/ta/{km,gk} and aosp/.../linaro/{km,gk}), we remove them
temporarily.

Also removed inclusion of ta/Android.mk in the main Android.mk file.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Signed-off-by: Victor Chong <victor.chong@linaro.org>